Reorganize into base read and base write query classes
